feat: implement FIBER_ROOT_TRACKER_SCRIPT for enhanced fiber commit tracking and name caching in React profiler
